:sparkles: Change the Structure to Repository
